+++ CHANGES.rst
@@ -8,6 +8,12 @@ Changelog
 - Update with new settings values from design.plone.contenttypes (version 3.0.0).
   [cekk]
 
+- [dev] Fix CI
+  [arsenico13]
+- **search-filters** endpoint now return also a list of searchable portal_types.
+  [cekk]
+- Disable some types from *types_not_searched*.
+  [cekk]
 
 1.1.0 (2021-03-24)
 ------------------